[YS-151] refactor: 컴포넌트 디렉토리 위치 수정
