+++ src/app/join/components/JoinCheckboxContainer/JoinCheckbox/JoinCheckbox.tsx
