+++ src/app/join/components/JoinCheckboxContainer/JoinCheckboxContainer.styles.ts
